\name{clobe-package}
\alias{clobe}
\docType{package}
\title{

    Qualitative Global Benchmark and Evaluation Toolkit for Formal Compuational Models

}

\description{
    
        Provides tools for the qualitative evaluation of formal computational models.
        Currently incorporates tools used in comparing model and human heterogeneity,
        see Dome and Wills (2023) <doi:10.31234/osf.io/ygmcj>.

Please cite the package in publications. Use \code{citation("clobe")}.

}

\keyword{computational modelling; parameter space partitioning; model evaluation}

\author{

    Lenard Dome

    Maintainer: Lenard Dome \email{lenarddome@gmail.com}

}

\references{

    Dome, L., & Wills, A. (2023, April 11). g-distance: On the comparison of model and human heterogeneity. PsyArxiv.

}